Updated AppContainer. Moved App. Updated App. Updated Navbar. Updated import/export.
