+++ client/src/modules/layout/index.js
@@ -1,5 +1,6 @@
+import App from './App';
 import HomePage from './HomePage';
 import Navbar from './Navbar';
 import Page404 from './Page404';
 
-export { HomePage, Navbar, Page404 };
+export { App, HomePage, Navbar, Page404 };
